requirements.txt
streamlit==1.56.0
websockets==16.0
requests==2.33.1
python-dotenv==1.2.2
langdetect==1.0.9